move binance downloader
